+++ src/blob/BlobConfiguration.ts
@@ -1,8 +1,6 @@
 import ConfigurationBase from "../common/ConfigurationBase";
 import {
-  DEFAULT_ACCESS_LOG_PATH,
   DEFAULT_BLOB_PERSISTENCE_PATH,
-  DEFAULT_DEBUG_LOG_PATH,
   DEFAULT_ENABLE_ACCESS_LOG,
   DEFAULT_ENABLE_DEBUG_LOG,
   DEFAULT_LOKI_DB_PATH,
@@ -29,15 +27,15 @@ export default class BlobConfiguration extends ConfigurationBase {
     public readonly dbPath: string = DEFAULT_LOKI_DB_PATH,
     public readonly persistencePath: string = DEFAULT_BLOB_PERSISTENCE_PATH,
     enableAccessLog: boolean = DEFAULT_ENABLE_ACCESS_LOG,
-    accessLogFilePath: string = DEFAULT_ACCESS_LOG_PATH,
+    accessLogWriteStream?: NodeJS.WritableStream,
     enableDebugLog: boolean = DEFAULT_ENABLE_DEBUG_LOG,
-    debugLogFilePath: string = DEFAULT_DEBUG_LOG_PATH
+    debugLogFilePath?: string
   ) {
     super(
       host,
       port,
       enableAccessLog,
-      accessLogFilePath,
+      accessLogWriteStream,
       enableDebugLog,
       debugLogFilePath
     );