+++ src/common/ConfigurationBase.ts
@@ -3,7 +3,7 @@ export default abstract class ConfigurationBase {
     public readonly host: string,
     public readonly port: number,
     public readonly enableAccessLog: boolean = false,
-    public readonly accessLogFilePath?: string,
+    public readonly accessLogWriteStream?: NodeJS.WritableStream,
     public readonly enableDebugLog: boolean = false,
     public readonly debugLogFilePath?: string
   ) {}