+++ tests/blob/apis/blob.test.ts
@@ -88,6 +88,7 @@ describe("BlobAPIs", () => {
   it("download with with default parameters", async () => {
     const result = await blobURL.download(Aborter.none, 0);
     assert.deepStrictEqual(await bodyToString(result, content.length), content);
+    assert.equal(result.contentRange, undefined);
   });
 
   it("download all parameters set", async () => {
@@ -95,6 +96,16 @@ describe("BlobAPIs", () => {
       rangeGetContentMD5: true
     });
     assert.deepStrictEqual(await bodyToString(result, 1), content[0]);
+    assert.equal(result.contentRange, `bytes 0-0/${content.length}`);
+  });
+
+  it("download entire with range", async () => {
+    const result = await blobURL.download(Aborter.none, 0, content.length);
+    assert.deepStrictEqual(await bodyToString(result, content.length), content);
+    assert.equal(
+      result.contentRange,
+      `bytes 0-${content.length - 1}/${content.length}`
+    );
   });
 
   it("delete", async () => {
@@ -378,5 +389,82 @@ describe("BlobAPIs", () => {
       result.contentDisposition,
       headers.blobContentDisposition
     );
+  });
+
+  it("Copy blob should work", async () => {
+    const sourceBlob = getUniqueName("blob");
+    const destBlob = getUniqueName("blob");
+
+    const sourceBlobURL = BlockBlobURL.fromContainerURL(
+      containerURL,
+      sourceBlob
+    );
+    const destBlobURL = BlockBlobURL.fromContainerURL(containerURL, destBlob);
+
+    const metadata = { key: "value" };
+    const blobHTTPHeaders = {
+      blobCacheControl: "blobCacheControl",
+      blobContentDisposition: "blobContentDisposition",
+      blobContentEncoding: "blobContentEncoding",
+      blobContentLanguage: "blobContentLanguage",
+      blobContentType: "blobContentType"
+    };
+
+    await sourceBlobURL.upload(Aborter.none, "hello", 5, {
+      metadata,
+      blobHTTPHeaders
+    });
+
+    await destBlobURL.startCopyFromURL(Aborter.none, sourceBlobURL.url);
+
+    const result = await destBlobURL.getProperties(Aborter.none);
+    assert.ok(result.date);
+    assert.deepStrictEqual(result.blobType, "BlockBlob");
+    assert.ok(result.lastModified);
+    assert.deepStrictEqual(result.metadata, metadata);
+    assert.deepStrictEqual(
+      result.cacheControl,
+      blobHTTPHeaders.blobCacheControl
+    );
+    assert.deepStrictEqual(result.contentType, blobHTTPHeaders.blobContentType);
+    assert.deepStrictEqual(
+      result.contentEncoding,
+      blobHTTPHeaders.blobContentEncoding
+    );
+    assert.deepStrictEqual(
+      result.contentLanguage,
+      blobHTTPHeaders.blobContentLanguage
+    );
+    assert.deepStrictEqual(
+      result.contentDisposition,
+      blobHTTPHeaders.blobContentDisposition
+    );
+  });
+
+  it("Copy blob should work to override metadata", async () => {
+    const sourceBlob = getUniqueName("blob");
+    const destBlob = getUniqueName("blob");
+
+    const sourceBlobURL = BlockBlobURL.fromContainerURL(
+      containerURL,
+      sourceBlob
+    );
+    const destBlobURL = BlockBlobURL.fromContainerURL(containerURL, destBlob);
+
+    const metadata = { key: "value" };
+    const metadata2 = { key: "value2" };
+    await sourceBlobURL.upload(Aborter.none, "hello", 5, {
+      metadata
+    });
+
+    await destBlobURL.startCopyFromURL(Aborter.none, sourceBlobURL.url, {
+      metadata: metadata2
+    });
+
+    const result = await destBlobURL.getProperties(Aborter.none);
+    assert.ok(result.date);
+    assert.deepStrictEqual(result.blobType, "BlockBlob");
+    assert.ok(result.lastModified);
+    assert.deepStrictEqual(result.metadata, metadata2);
   });
 });