+++ ChangeLog.md
@@ -2,6 +2,16 @@
 
 > Note. This file includes changes after 3.0.0-preview. For legacy Azurite changes, please goto GitHub [releases](https://github.com/Azure/Azurite/releases).
 
+2019.06 Version 3.1.0-preview
+
+- Integrated Azurite with Visual Studio Code as an extension.
+- Added Visual Studio Code extension usage guidelines.
+- Added Dockerfile and usage descriptions.
+- Fixed an authentication issue when copy blob to override an existing blob with SAS.
+- Return 404 for copy blob operation when source blob doesn't exist.
+- Fixed an issue that metadata doesn't get copied when copy blob.
+- Fixed GetBlockBlob missing Content-Range header
+
 2019.05 Version 3.0.0-preview
 
 - Initial Release of Azurite V3.